Adding financial year

+++ db_scripts/tables/tbl_vendcyc.sql
@@ -12,7 +12,7 @@ create table vendcyc
   total_amount_ordered numeric,
   amount_ordered numeric,
   amount_funded numeric,
-  amount_invoiced_funded integer,
+  amount_invoiced_funded numeric,
   total_paid numeric, 
   deposit_encumbrance_margin integer,
   ordered integer,
